+++ cmd/mastiffgen/internal/scaffold/templates/init/go.mod.tmpl
@@ -5,7 +5,7 @@ go 1.24.4
 require (
 	github.com/alicebob/miniredis/v2 v2.35.0
 	github.com/go-redis/redis/v7 v7.4.1
-	github.com/hewen/mastiff-go v0.0.0-20250728041216-91a44872d1a1
+	github.com/hewen/mastiff-go v0.0.0-20250822035448-b697515cbbe8
 	github.com/mitchellh/go-homedir v1.1.0
 	github.com/panjf2000/gnet/v2 v2.9.1
 	github.com/spf13/cobra v1.9.1
