+++ cmd/mastiffgen/internal/scaffold/templates/init/go.sum.tmpl
@@ -200,8 +200,8 @@ github.com/hashicorp/logutils v1.0.0/go.mod h1:QIAnNjmIWmVIIkWDTG1z5v++HQmx9WQRO
 github.com/hashicorp/mdns v1.0.0/go.mod h1:tL+uN++7HEJ6SQLQ2/p+z2pH24WQKWjBPkE0mNTz8vQ=
 github.com/hashicorp/memberlist v0.1.3/go.mod h1:ajVTdAv/9Im8oMAAj5G31PhhMCZJV2pPBoIllUwCN7I=
 github.com/hashicorp/serf v0.8.2/go.mod h1:6hOLApaqBFA1NXqRQAsxw9QxuDEvNxSQRwA/JwenrHc=
-github.com/hewen/mastiff-go v0.0.0-20250728041216-91a44872d1a1 h1:J9GsqNQliq3VEwp7xGOAqpyAKFDctyli1N980nJsz8M=
-github.com/hewen/mastiff-go v0.0.0-20250728041216-91a44872d1a1/go.mod h1:H741ZqJd2izdCCsilSiu4c8jVO3GapWgCyYxDt34V9M=
+github.com/hewen/mastiff-go v0.0.0-20250822035448-b697515cbbe8 h1:Wrs8qfgCSw83mnwp665wI03bD3x3oj01FeIAe/+21Rc=
+github.com/hewen/mastiff-go v0.0.0-20250822035448-b697515cbbe8/go.mod h1:H741ZqJd2izdCCsilSiu4c8jVO3GapWgCyYxDt34V9M=
 github.com/hpcloud/tail v1.0.0 h1:nfCOvKYfkgYP8hkirhJocXT2+zOD8yUNjXaWfTlyFKI=
 github.com/hpcloud/tail v1.0.0/go.mod h1:ab1qPbhIpdTxEkNHXyeSf5vhxWSCs/tWer42PpOxQnU=
 github.com/hudl/fargo v1.3.0/go.mod h1:y3CKSmjA+wD2gak7sUSXTAoopbhU08POFhmITJgmKTg=
