+++ middleware/auth/httpx.go
@@ -5,7 +5,7 @@ import (
 	"net/http"
 
 	"github.com/hewen/mastiff-go/config/middlewareconf/authconf"
-	"github.com/hewen/mastiff-go/internal/contextkeys"
+	"github.com/hewen/mastiff-go/pkg/contextkeys"
 	"github.com/hewen/mastiff-go/server/httpx/unicontext"
 )
