+++ middleware/ratelimit/limit.go
@@ -9,7 +9,7 @@ import (
 	"time"
 
 	"github.com/hewen/mastiff-go/config/middlewareconf/ratelimitconf"
-	"github.com/hewen/mastiff-go/internal/contextkeys"
+	"github.com/hewen/mastiff-go/pkg/contextkeys"
 	"github.com/hewen/mastiff-go/server/httpx/unicontext"
 	"golang.org/x/time/rate"
 	"google.golang.org/grpc/peer"
