+++ store/redishook.go
@@ -6,8 +6,8 @@ import (
 	"time"
 
 	"github.com/go-redis/redis/v7"
-	"github.com/hewen/mastiff-go/internal/contextkeys"
 	"github.com/hewen/mastiff-go/logger"
+	"github.com/hewen/mastiff-go/pkg/contextkeys"
 	"github.com/hewen/mastiff-go/pkg/util"
 )
